+++ apps/cli/mod.ts
@@ -1,9 +1,6 @@
-export { buildWithDocker, buildWithDockerAndValidate } from "@/utils"
-export type { BuildResult } from "@/types"
 export {
 	verifyCustomUiCommand,
 	verifyTemplateUiCommand,
 	verifySingleProject,
 	verifyCmd
 } from "@/commands/verify"
-export { buildCmd } from "@/commands/build"