+++ packages/api/src/api/server-types.ts
@@ -6,7 +6,7 @@ import type {
 	IOPayload,
 	SpawnOptions
 } from "tauri-plugin-shellx-api"
-import type { DenoRunConfig, IUiIframe } from "./client"
+import type { DenoRunConfig, IUiCustom } from "./client"
 
 export type IShellServer = IShellServer1 & {
 	denoExecute(
@@ -25,13 +25,13 @@ export type IShellServer = IShellServer1 & {
 }
 
 // This will be implemented in the @kksh/api package
-export type IUiIframeServer1 = Pick<
-	IUiIframe,
+export type IUiCustomServer1 = Pick<
+	IUiCustom,
 	"startDragging" | "toggleMaximize" | "internalToggleMaximize"
 >
 // This interface will be implemented in iframe-ext.vue where iframe is loaded and API is exposed
 // because these API dependes on the context of the page
-export type IUiIframeServer2 = Omit<
-	IUiIframe,
+export type IUiCustomServer2 = Omit<
+	IUiCustom,
 	"registerDragRegion" | "internalToggleMaximize" | "toggleMaximize" | "startDragging"
 >