+++ packages/api/src/headless/ext.ts
@@ -1,4 +1,4 @@
-export abstract class HeadlessWorkerExtension {
+export abstract class HeadlessCommand {
 	/**
 	 * Load the extension. Initialize the extension.
 	 * Will be called once when the extension is first loaded.