+++ packages/api/src/ui/custom/index.ts
@@ -1,9 +1,5 @@
-/* -------------------------------------------------------------------------- */
-/*                                     API                                    */
-/* -------------------------------------------------------------------------- */
-// import { wrap, type Endpoint, type Remote } from "@huakunshen/comlink"
-// import { RPCChannel, WorkerChildIO, type DestroyableIoInterface } from "kkrpc/browser"
-import { RPCChannel, WorkerChildIO, type DestroyableIoInterface } from "kkrpc/browser"
+// import { windowEndpoint, wrap, type Remote } from "@huakunshen/comlink"
+import { IframeChildIO, RPCChannel, type DestroyableIoInterface } from "kkrpc/browser"
 import type {
 	IClipboard,
 	IDialog,
@@ -22,7 +18,6 @@ import type {
 import {
 	constructFetchAPI,
 	// constructPathAPI,
-	// constructShellAPI,
 	constructUpdownloadAPI
 } from "tauri-api-adapter/client"
 import type {
@@ -36,29 +31,20 @@ import type {
 	ISecurity,
 	ISystem,
 	IToast,
+	IUiCustom,
 	IUtils
 } from "../../api/client"
 import { constructEventAPI } from "../../api/event"
+import { constructIframeUiAPI } from "../../api/iframe-ui"
 import { constructPathAPI } from "../../api/path"
 import type { IShellServer } from "../../api/server-types"
 import { constructShellAPI } from "../../api/shell"
-import { constructToastAPI } from "../../api/toast"
-import type { FormSchema, ListSchema, MarkdownSchema } from "../../models"
-import type { IComponent } from "./components"
-import type { WorkerExtension } from "./ext"
 
-export interface IUiWorker {
-	render: (view: IComponent<ListSchema.List | FormSchema.Form | MarkdownSchema>) => Promise<void>
-	goBack: () => Promise<void>
-	showLoadingBar: (loading: boolean) => Promise<void>
-	setScrollLoading: (loading: boolean) => Promise<void>
-	setSearchTerm: (term: string) => Promise<void>
-	setSearchBarPlaceholder: (placeholder: string) => Promise<void>
-	setProgressBar: (progress: number | null) => Promise<void>
-}
-
+export { type IUiCustom } from "../../api/client"
 // export { expose, wrap } from "@huakunshen/comlink"
-export { WorkerExtension } from "./ext"
+// export { type IDbServer } from "../server/db"
+export { type IUiCustomServer2, type IUiCustomServer1 } from "../../api/server-types"
+
 /**
  * For the APIs annotated with "inherit from tauri-api-adapter", they inherit the client API completely from tauri-api-adapter
  * There may be server API changes for them, but the client API can be inherited
@@ -82,34 +68,27 @@ type API = {
 	updownload: IUpdownload // inherit from tauri-api-adapter
 	sysInfo: ISystemInfo // inherit from tauri-api-adapter
 	network: INetwork // inherit from tauri-api-adapter
-	workerUi: IUiWorker // for kunkun
-	security: ISecurity // for kunkun
+	iframeUi: IUiCustom // for kunkun
 	utils: IUtils // for kunkun
+	security: ISecurity // for kunkun
 	app: IApp
 }
-
-/* -------------------------------------------------------------------------- */
-/*                                   Expose                                   */
-/* -------------------------------------------------------------------------- */
-const io = new WorkerChildIO()
+// export const api = wrap(windowEndpoint(globalThis.parent)) as unknown as API
+const io = new IframeChildIO()
 const rpc = new RPCChannel<{}, API, DestroyableIoInterface>(io, {})
 export const api = rpc.getAPI()
-
-export function expose(api: WorkerExtension) {
-	rpc.expose(api)
-}
 
 export const event = constructEventAPI(api.event) // this is different from event api from tauri-api-adapter
 export const fetch = constructFetchAPI(api.fetch)
 export const path = constructPathAPI(api.path)
 export const shell = constructShellAPI(api.shell)
-export const toast = constructToastAPI(api.toast)
 export const updownload = constructUpdownloadAPI(api.updownload)
+export const ui = constructIframeUiAPI(api)
 export const {
 	db,
-	kv,
 	os,
 	clipboard,
+	kv,
 	dialog,
 	fs,
 	log,
@@ -117,24 +96,9 @@ export const {
 	sysInfo,
 	network,
 	system,
-	open,
+	toast,
 	utils,
-	app,
-	security,
-	workerUi: ui
+	open,
+	app
 } = api
 export { Child, RPCChannel, Command, DenoCommand } from "../../api/shell"
-/* -------------------------------------------------------------------------- */
-/*                             UI Component Schema                            */
-/* -------------------------------------------------------------------------- */
-export { type IComponent } from "./components"
-export * as ListSchema from "./schema/list"
-export * as FormSchema from "./schema/form"
-export { Markdown as MarkdownSchema } from "./schema/markdown"
-
-export { List, Action, Form, Markdown } from "./components"
-export * from "../../models/styles"
-export { Icon } from "./components/icon"
-export { IconEnum, IconType, IconNode } from "../../models/icon"
-export * as schema from "./schema"
-export { NodeName, NodeNameEnum, FormNodeName, FormNodeNameEnum } from "../../models/constants"