+++ packages/extensions/demo-worker-template-ext/src/headless.ts
@@ -1,6 +1,6 @@
-import { expose, HeadlessWorkerExtension, toast } from "@kksh/api/headless"
+import { expose, HeadlessCommand, toast } from "@kksh/api/headless"
 
-class DemoHeadlessExt extends HeadlessWorkerExtension {
+class DemoHeadlessExt extends HeadlessCommand {
 	load(): Promise<void> {
 		console.log("Demo Headless Extension Loaded")
 		toast.info("Demo Headless Extension Loaded")