+++ packages/extensions/demo-worker-template-ext/CHANGELOG.md
@@ -1,5 +1,19 @@
 # demo-template-extension
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.1
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.0
+
 ## 0.0.6
 
 ### Patch Changes