+++ packages/templates/template-ext-svelte/CHANGELOG.md
@@ -1,5 +1,19 @@
 # template-ext-svelte
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.0
+
 ## 0.0.5
 
 ### Patch Changes