+++ packages/templates/template-ext-vue/CHANGELOG.md
@@ -1,5 +1,19 @@
 # template-ext-vue
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.1
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @kksh/api@0.1.0
+
 ## 0.0.3
 
 ### Patch Changes